project/myapp/requirements/dev.txt
#Requirements
ckeditor
django-import-export
django-admin-interface